fix: address PR review issues for link_toc job
- Move LoadTocEntries to common package, removing duplicates from
  link_toc and process_book
- Add TocEntries field to BookState for caching during LoadBook
- Fix grep_text cluster gap comment to match code (<=3 not <=2)
- Add logging to all nil-return paths for better debuggability
