+++ internal/jobs/link_toc/link_toc.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	toc_entry_finder "github.com/jackzampolin/shelf/internal/agents/toc_entry_finder"
 	"github.com/jackzampolin/shelf/internal/jobs"
 	"github.com/jackzampolin/shelf/internal/jobs/common"
 	ljob "github.com/jackzampolin/shelf/internal/jobs/link_toc/job"
@@ -55,11 +54,8 @@ func NewJob(ctx context.Context, cfg Config, bookID string) (jobs.Job, error) {
 		return nil, fmt.Errorf("ToC extraction not complete - cannot link entries")
 	}
 
-	// Load TocEntry records
-	entries, err := LoadTocEntries(ctx, result.TocDocID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load ToC entries: %w", err)
-	}
+	// TocEntries are loaded during LoadBook when TocExtract is complete
+	entries := result.Book.GetTocEntries()
 
 	logger := svcctx.LoggerFrom(ctx)
 	if logger != nil {
@@ -70,7 +66,7 @@ func NewJob(ctx context.Context, cfg Config, bookID string) (jobs.Job, error) {
 			"entries_count", len(entries))
 	}
 
-	return ljob.NewFromLoadResult(result, entries), nil
+	return ljob.NewFromLoadResult(result), nil
 }
 
 // JobFactory returns a factory function for recreating jobs from stored metadata.
@@ -78,86 +74,4 @@ func JobFactory(cfg Config) jobs.JobFactory {
 	return common.MakeJobFactory(func(ctx context.Context, bookID string) (jobs.Job, error) {
 		return NewJob(ctx, cfg, bookID)
 	})
-}
-
-// LoadTocEntries loads all TocEntry records for a ToC.
-func LoadTocEntries(ctx context.Context, tocDocID string) ([]*toc_entry_finder.TocEntry, error) {
-	if tocDocID == "" {
-		return nil, fmt.Errorf("ToC document ID is required")
-	}
-
-	defraClient := svcctx.DefraClientFrom(ctx)
-	if defraClient == nil {
-		return nil, fmt.Errorf("defra client not in context")
-	}
-
-	query := fmt.Sprintf(`{
-		TocEntry(filter: {toc_id: {_eq: "%s"}}, order: {sort_order: ASC}) {
-			_docID
-			entry_number
-			title
-			level
-			level_name
-			printed_page_number
-			sort_order
-			actual_page {
-				_docID
-			}
-		}
-	}`, tocDocID)
-
-	resp, err := defraClient.Execute(ctx, query, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	rawEntries, ok := resp.Data["TocEntry"].([]any)
-	if !ok {
-		return nil, nil // No entries
-	}
-
-	var entries []*toc_entry_finder.TocEntry
-	for _, e := range rawEntries {
-		entry, ok := e.(map[string]any)
-		if !ok {
-			continue
-		}
-
-		// Skip entries that already have actual_page linked
-		if actualPage, ok := entry["actual_page"].(map[string]any); ok {
-			if _, hasDoc := actualPage["_docID"]; hasDoc {
-				continue // Already linked
-			}
-		}
-
-		te := &toc_entry_finder.TocEntry{}
-
-		if docID, ok := entry["_docID"].(string); ok {
-			te.DocID = docID
-		}
-		if entryNum, ok := entry["entry_number"].(string); ok {
-			te.EntryNumber = entryNum
-		}
-		if title, ok := entry["title"].(string); ok {
-			te.Title = title
-		}
-		if level, ok := entry["level"].(float64); ok {
-			te.Level = int(level)
-		}
-		if levelName, ok := entry["level_name"].(string); ok {
-			te.LevelName = levelName
-		}
-		if printedPage, ok := entry["printed_page_number"].(string); ok {
-			te.PrintedPageNumber = printedPage
-		}
-		if sortOrder, ok := entry["sort_order"].(float64); ok {
-			te.SortOrder = int(sortOrder)
-		}
-
-		if te.DocID != "" {
-			entries = append(entries, te)
-		}
-	}
-
-	return entries, nil
 }
